requirements.txt
# requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-socketio>=5.8.0
python-dotenv>=1.0.0
groq>=0.4.0
requests>=2.31.0
sqlalchemy>=2.0.0
python-docx>=0.8.11
reportlab>=4.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
mcp>=0.1.0